Move apportion to lambda_schedule.py
* Add some additional assertions
